Added new permission types, added new decorators to get authUser and permissionMeta, updated services and controllers to be compatible with new features
- feat: global, scoped and contextual permission types; decorators for obtaining AuthUser and PermissionMeta; added findPermissionByPrefix function to obtain permissions matching a prefix
- fix: added scope checks where missing
- ref: changed types of several permissions

+++ backend/src/modules/character/commands/delete-character/delete-character.handler.ts
@@ -2,9 +2,10 @@ import { CommandHandler } from '@nestjs/cqrs';
 import { ICommandHandler } from '@nestjs/cqrs';
 import { DeleteCharacterCommand } from './delete-character.command.js';
 import { CharacterRepository } from '../../character.repository.js';
-import { NotFoundError } from '../../../../shared/errors/app.error.js';
+import { ForbiddenError, NotFoundError } from '../../../../shared/errors/app.error.js';
 import { UnitOfWork } from '../../../../core/database/unit-of-work.js';
 import { OwnershipService } from '../../../../shared/modules/ownership/ownership.service.js';
+import { PermissionType } from '@project/shared';
 
 @CommandHandler(DeleteCharacterCommand)
 export class DeleteCharacterHandler implements ICommandHandler<DeleteCharacterCommand> {
@@ -16,7 +17,11 @@ export class DeleteCharacterHandler implements ICommandHandler<DeleteCharacterCo
   async execute(command: DeleteCharacterCommand) {
     const character = await this.characterRepository.findById(command.characterId);
     if (!character) throw new NotFoundError('Character');
-    this.ownershipService.checkCharacter(character, command.userId, command.scope);
+
+    if (command.permissionMeta.type !== PermissionType.SCOPED) throw new ForbiddenError();
+    const scopes = command.permissionMeta.scopes;
+    if (!scopes.includes('any')) this.ownershipService.checkCharacter(character, command.userId);
+
     await this.characterRepository.delete(character);
     await this.uow.saveChanges();
   }
